core: Async cleanup
